+++ src/main/java/ru/vyarus/dropwizard/guice/module/installer/feature/ResourceInstaller.java
@@ -3,19 +3,19 @@ package ru.vyarus.dropwizard.guice.module.installer.feature;
 
 import io.dropwizard.setup.Environment;
 import ru.vyarus.dropwizard.guice.module.installer.FeatureInstaller;
-import ru.vyarus.dropwizard.guice.module.installer.install.TypeInstaller;
+import ru.vyarus.dropwizard.guice.module.installer.install.InstanceInstaller;
 import ru.vyarus.dropwizard.guice.module.installer.util.FeatureUtils;
 
 import javax.ws.rs.Path;
 
 /**
  * Jersey resource installer.
- * Search classes annotated with {@code @Path}. Installers mechanism will register bean in context and
- * jersey-guice will handle registration into dropwizard jersey context (no need for special installation logic)
- * <p>NOTE: if you will nor annotate resource with @Singleton, it will be used as prototype and
- * instantiated on each request!</p>
+ * Search classes annotated with {@code @Path}. Directly register instance in jersey context to force singleton.
+ * If we register it by type, then we could use prototype beans (resource instance created on each request),
+ * which will lead to performance overhead. To ovoid misuse, singleton resources are forced. Override installer
+ * if you really need prototype resources.
  */
-public class ResourceInstaller implements FeatureInstaller<Object>, TypeInstaller<Object> {
+public class ResourceInstaller implements FeatureInstaller<Object>, InstanceInstaller<Object> {
 
     @Override
     public boolean matches(final Class<?> type) {
@@ -23,7 +23,7 @@ public class ResourceInstaller implements FeatureInstaller<Object>, TypeInstalle
     }
 
     @Override
-    public void install(final Environment environment, final Class<Object> type) {
-        environment.jersey().register(type);
+    public void install(final Environment environment, final Object instance) {
+        environment.jersey().register(instance);
     }
 }