+++ src/test/groovy/ru/vyarus/dropwizard/guice/support/feature/DummyService.groovy
@@ -1,12 +1,12 @@
 package ru.vyarus.dropwizard.guice.support.feature
 
-import ru.vyarus.dropwizard.guice.module.installer.feature.eager.Eager
+import ru.vyarus.dropwizard.guice.module.installer.feature.eager.EagerSingleton
 
 /**
  * @author Vyacheslav Rusakov 
  * @since 03.09.2014
  */
-@Eager
+@EagerSingleton
 class DummyService {
 
     public String hey() {