+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 * Installers may choose now from three types of installation (binding, type or instance) or combine them.
 * Add PluginInstaller: shortcut for multibindings mechanism
 * Updated guice (4.0.beta4 -> 4.0.beta5)
+* Force singleton for resources
+* @Eager renamed to @EagerSingleton and now forces singleton scope for bean
 
 ### 0.9.0 (2014-09-05)